+++ src/renderer-app/src/AppRoutes.tsx
@@ -3,7 +3,7 @@ import { HashRouter } from "react-router-dom";
 import { Route, Switch } from "react-router";
 import { message } from "antd";
 import WhiteboardCreatorPage from "./WhiteboardCreatorPage";
-import IndexPage from "./IndexPage";
+// import IndexPage from "./IndexPage";
 import WhiteboardPage from "./WhiteboardPage";
 import ReplayPage from "./ReplayPage";
 import JoinPage from "./JoinPage";
@@ -17,6 +17,7 @@ import UserInfoPage from "./UserInfoPage";
 import UserSetPage from "./UserSetPage";
 import UserScheduledPage from "./UserScheduledPage";
 import RoomDetailPage from "./RoomDetailPage";
+import LoginPage from "./LoginPage";
 
 export class AppRoutes extends React.Component<{}, {}> {
     public componentDidCatch(error: any): void {
@@ -43,7 +44,7 @@ export class AppRoutes extends React.Component<{}, {}> {
                     <Route exact path="/create/" component={CreatePage} />
                     <Route exact path="/storage/" component={Storage} />
                     <Route exact path="/history/" component={HistoryPage} />
-                    <Route exact path="/" component={IndexPage} />
+                    <Route exact path="/" component={LoginPage} />
                     <Route exact path="/user/" component={UserIndexPage} />
                     <Route exact path="/user/scheduled/" component={UserScheduledPage} />
                     <Route exact path="/user/room/" component={RoomDetailPage} />
