+++ src/renderer-app/src/UserSetPage.less
@@ -4,9 +4,6 @@
     width: 240px;
 
     .ant-menu-item {
-        &:first-child {
-            margin-top: 24px;
-        }
         a:hover {
             color: #444E60;
         }
@@ -23,6 +20,10 @@
         color: #444E60;
     }
     
+}
+
+.menu-item-container.ant-menu {
+    padding-top: 24px;
 }
 
 .setting-content-container {
